backend/requirements.txt
fastapi
uvicorn
python-dotenv
firebase-admin
google-genai
google-auth
requests
